Add option to keep Ray clusters after a job has finished (#794)
* add shutdown delay for ray clusters to gateway


* fix lint


* remove delayed remove option


* lint


* add label for easy delete


* I wish you had... more time


* debug log


---------

+++ gateway/api/management/commands/free_resources.py
@@ -6,6 +6,7 @@ from django.core.management.base import BaseCommand
 
 from api.models import ComputeResource, Job
 from api.ray import kill_ray_cluster
+from main import settings as config
 
 
 logger = logging.getLogger("commands")
@@ -27,6 +28,13 @@ class Command(BaseCommand):
 
             # only kill cluster if not in local mode and no jobs are running there
             if len(alive_jobs) == 0 and not settings.RAY_CLUSTER_MODE.get("local"):
+                if config.RAY_CLUSTER_NO_DELETE_ON_COMPLETE:
+                    logger.debug(
+                        "RAY_CLUSTER_NO_DELETE_ON_COMPLETE is enabled, "
+                        + "so cluster [%s] will not be removed",
+                        compute_resource.title,
+                    )
+                    return
                 kill_ray_cluster(compute_resource.title)
                 compute_resource.delete()
                 counter += 1
